Improved streaming behaviour
On highly compressible data streams Z_NO_FLUSH means no data will be output in chunks of an optimal size. This breaks the expected streaming behavior so for example if sending to an HTTP socket the client will timeout and drop the connection due to the data compression too well! Changing to Z_SYNC_FLUSH causes a minor synchronization point which barely affects the compression ration at all.